Move common module to samples project.
