+++ samples/kotlin-multiplatform-common/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/kotlin/common/Sample.kt
@@ -1,4 +1,4 @@
-package org.jlleitschuh.gradle.ktlint.sample.kotlin
+package org.jlleitschuh.gradle.ktlint.sample.kotlin.common
 
 /**
  * Each platform must implement a Sample class matching this signature.
